migrated config to common
