+++ src/weathergen/run_train.py
@@ -18,7 +18,7 @@ import traceback
 from pathlib import Path
 
 import weathergen.utils.cli as cli
-import weathergen.utils.config as config
+import weathergen.common.config as config
 from weathergen.train.trainer import Trainer
 from weathergen.utils.logger import init_loggers
