+++ src/weathergen/train/trainer.py
@@ -27,12 +27,13 @@ from torch.distributed.fsdp.wrap import (
 )
 
 import weathergen.train.loss as losses
-import weathergen.utils.config as config
+import weathergen.common.config as config
 from weathergen.datasets.multi_stream_data_sampler import MultiStreamDataSampler
 from weathergen.model.model import Model, ModelParams
 from weathergen.train.lr_scheduler import LearningRateScheduler
 from weathergen.train.trainer_base import TrainerBase
-from weathergen.utils.config import Config, get_dtype
+from weathergen.utils.utils import get_dtype
+from weathergen.common.config import Config
 from weathergen.utils.distributed import is_root
 from weathergen.utils.train_logger import TRAIN, VAL, TrainLogger
 from weathergen.utils.validation_io import write_output
