+++ src/weathergen/utils/validation_io.py
@@ -10,7 +10,7 @@
 import logging
 
 import weathergen.common.io as io
-import weathergen.utils.config as config
+import weathergen.common.config as config
 
 _logger = logging.getLogger(__name__)
