+++ tests/test_config.py
@@ -4,7 +4,7 @@ import tempfile
 import pytest
 from omegaconf import OmegaConf
 
-import weathergen.utils.config as config
+import weathergen.common.config as config
 
 TEST_RUN_ID = "test123"
 SECRET_COMPONENT = "53CR3T"
@@ -217,7 +217,7 @@ def test_from_cli(options, cf):
     ],
 )
 def test_set_run_id(config_fresh, run_id, reuse, expected, mocker):
-    mocker.patch("weathergen.utils.config.get_run_id", return_value="generated")
+    mocker.patch("weathergen.common.config.get_run_id", return_value="generated")
 
     config_fresh = config.set_run_id(config_fresh, run_id, reuse)
